Changed homepage

+++ src/Controller/ArticlesController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Article;
 use App\Repository\ArticleRepository;
+use App\Repository\UserRepository;
 use App\Form\ArticleCreateFormType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
@@ -15,20 +16,21 @@ use App\Service\FileUploader;
 class ArticlesController extends AbstractController
 {
     private $articleRepository;
+    private $userRepository;
 
-    public function __construct(ArticleRepository $articleRepository)
+    public function __construct(ArticleRepository $articleRepository, UserRepository $userRepository, EntityManagerInterface $entityManager)
     {
         $this->articleRepository = $articleRepository;
+        $this->userRepository = $userRepository;
+        $this->entityManager = $entityManager;
     }
 
     #[Route('/', methods: ['GET'], name: 'app_articles')]
     public function index(): Response
     {
-        $articles = $this->articleRepository->findAll();
-        // dd($articles);
-
+        $articles = $this->articleRepository->findBy([], ['created_at' => 'DESC']);
+        
         return $this->render('articles/index.html.twig', [
-            'controller_name' => 'ArticlesController',
             'articles' => $articles
         ]);
     }
@@ -43,29 +45,34 @@ class ArticlesController extends AbstractController
         ]);
     }
 
-    #[Route('/article/create', name: 'app_article_create')]
-    public function create(Request $request, EntityManagerInterface $entityManager, FileUploader $fileUploader): Response
+    #[Route('/article/create/{userId<\d+>}', name: 'app_article_create')]
+    public function create($userId, Request $request, EntityManagerInterface $entityManager, FileUploader $fileUploader): Response
     {
-        $article = new Article();
-        $form = $this->createForm(ArticleCreateFormType::class, $article);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $articleImageFile = $form->get('image_path')->getData();
-            $filePath = $fileUploader->uploadImageToCloudinary($articleImageFile, 'images');
-            $article->setImagePath($filePath);
-            $article->setCreatedAt(new \DateTime("now"));
-            $article->setAuthor($this->getUser());
+        if ( $userId == $this->getUser()->getId() || $this->isGranted('ROLE_ADMIN') ) {
+            $article = new Article();
+            $form = $this->createForm(ArticleCreateFormType::class, $article);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $articleImageFile = $form->get('image_path')->getData();
+                $filePath = $fileUploader->uploadImageToCloudinary($articleImageFile, 'images');
+                $article->setImagePath($filePath);
+                $article->setCreatedAt(new \DateTime("now"));
+                $article->setAuthor($this->userRepository->find($userId));
+                
+                $entityManager->persist($article);
+                $entityManager->flush();
             
-            $entityManager->persist($article);
-            $entityManager->flush();
-        
-            return $this->redirectToRoute('app_profile', ['id' => $this->getUser()->getId()]);
-        }
+                return $this->redirectToRoute('app_profile_articles', ['id' => $userId]);
+            }
 
-        return $this->render('articles/create.html.twig', [
-            'articleCreateForm' => $form->createView()
-        ]);
+            return $this->render('articles/create.html.twig', [
+                'articleCreateForm' => $form->createView()
+            ]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 
     #[Route('/article/edit/{id<\d+>}', name: 'app_article_edit')]
@@ -74,7 +81,7 @@ class ArticlesController extends AbstractController
 
         $article = $this->articleRepository->find($id);
 
-        if ( $article->getAuthor()->getId() == $this->getUser()->getId() ) {
+        if ( $article->getAuthor()->getId() == $this->getUser()->getId() || $this->isGranted('ROLE_ADMIN') ) {
             $form = $this->createForm(ArticleCreateFormType::class, $article);
             $form->handleRequest($request);
 
@@ -107,10 +114,15 @@ class ArticlesController extends AbstractController
     public function delete($id, EntityManagerInterface $entityManager, FileUploader $fileUploader): Response
     {
         $article = $this->articleRepository->find($id);
-        $fileUploader->destroyImageCloudinary($article->getImagePath());
-        $entityManager->remove($article);
-        $entityManager->flush();
+        
+        if ( $article->getAuthor()->getId() == $this->getUser()->getId() || $this->isGranted('ROLE_ADMIN') ) {
+            $fileUploader->destroyImageCloudinary($article->getImagePath());
+            $entityManager->remove($article);
+            $entityManager->flush();
 
-        return $this->redirectToRoute('app_profile_articles', ['id' => $this->getUser()->getId()]);
+            return $this->redirectToRoute('app_profile_articles', ['id' => $this->getUser()->getId()]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 }
